refactor: move importer logic to new class
The new Importer class allows configuration to be stored in instance variables instead of having to be passed along a chain of function calls. Only code relating to the click cli remains in feed_ursus.py
