+++ tests/test_integration.py
@@ -10,8 +10,6 @@ from pysolr import Solr  # type: ignore
 
 from feed_ursus import feed_ursus
 
-feed_ursus.mapper = importlib.import_module("feed_ursus.mapper.sinai")
-
 
 SOLR_URL = os.getenv("SOLR_URL", "http://localhost:8983/solr/californica")
